Switch to Net::SMTP::TLS.
No longer use Net::SMTP or Net::STMP_auth, as ::TLS can handle all of that. It's an independent implementation that does not inherit from Net::STMP, but since it does all that stuff it seems like the better choice at this point (but watch [RT #40478](https://rt.cpan.org/Ticket/Display.html?id=40478) for changes to Net::SMTP).

This change also deprecates the `--smtp-authtype` option, which is now a no-op, and adds `--smtp-tls` for obvious reasons.

Based on a patch from Marino Miculan. Resolves RT#40188.