Account for non-glob stash entries
In bleadperl (5.28-to-be), a sub may be stored as a simple code ref in
the stash.  Even in existing stable releases of perl, stash entries
are not always typeglobs.  Constants created by ‘use constant’, for
instance, are stored as scalar refs.  This patch takes all these into
account, but does not add any tests, since I am not sure how to trig-
ger this code path with a constant.  Nevertheless, before this, the
tests failed with bleadperl.